fix: add unit of measure and precision for spinedits

+++ VerticalWarehouses/Panel PC UI/Ferretto.VW.App.Services/Interfaces/IWmsDataProvider.cs
@@ -10,7 +10,7 @@ namespace Ferretto.VW.App.Services
 
         Task<string> GetItemImagePathAsync(int itemId);
 
-        Task PickAsync(int itemId, int requestedQuantity);
+        Task PickAsync(int itemId, double requestedQuantity);
 
         #endregion
     }
